+++ test/unit/event-reduce.test.ts
@@ -46,10 +46,17 @@ describe('event-reduce.test.js', () => {
         res1: RxDocument<RxDocType>[],
         res2: RxDocument<RxDocType>[]
     ) {
-        assert.deepStrictEqual(
-            res1.map(d => d.primary),
-            res2.map(d => d.primary)
-        );
+        const keys1 = res1.map(d => d.primary);
+        const keys2 = res2.map(d => d.primary);
+
+        try {
+            assert.deepStrictEqual(keys1, keys2);
+        } catch (err) {
+            console.error('ensureResultsEqual() keys not equal');
+            console.dir(keys1);
+            console.dir(keys2);
+            throw err;
+        }
         assert.deepStrictEqual(
             res1.map(d => d.toJSON()),
             res2.map(d => d.toJSON())
@@ -60,13 +67,14 @@ describe('event-reduce.test.js', () => {
         col2: RxCollection<RxDocType>,
         queries: MangoQuery<RxDocType>[]
     ) {
-        await Promise.all(
-            queries.map(async (query) => {
-                const res1 = await col1.find(query).exec();
-                const res2 = await col2.find(query).exec();
-                ensureResultsEqual(res1, res2);
-            })
-        );
+        for (const query of queries) {
+            const [res1, res2] = await Promise.all([
+                col1.find(query).exec(),
+                col2.find(query).exec()
+            ]);
+            console.dir(query);
+            ensureResultsEqual(res1, res2);
+        }
     }
 
     it('should have the same results on given data', async () => {
@@ -142,7 +150,11 @@ describe('event-reduce.test.js', () => {
 
         const queries: MangoQuery<any>[] = [
             { selector: { age: { '$gt': 10 } }, sort: [{ passportId: 'asc' }] },
-            { selector: { firstName: { $eq: 'Freeman' } }, sort: [{ passportId: 'asc' }] }
+            { selector: { firstName: { $eq: 'Freeman' } }, sort: [{ passportId: 'asc' }] },
+            {
+                selector: {},
+                sort: [{ firstName: 'asc' }]
+            }
         ];
 
 
@@ -157,14 +169,14 @@ describe('event-reduce.test.js', () => {
 
         const writeData = [
             {
-                passportId: 's90j6hhznefj',
-                firstName: 'Freeman',
+                passportId: 's90j6hhznefj-bbbbb',
+                firstName: 'bbbbb',
                 lastName: 'Rogahn',
                 age: 25
             },
             {
-                passportId: '6eu7byz49iq9',
-                firstName: 'Eugenia',
+                passportId: '6eu7byz49iq9-aaaa',
+                firstName: 'aaaaa',
                 lastName: 'Dare',
                 age: 16
             }
@@ -182,6 +194,20 @@ describe('event-reduce.test.js', () => {
             queries
         );
 
+
+        const insertForSortTest = {
+            passportId: 'for-sort-00000',
+            firstName: '00000',
+            lastName: 'Rogahn2',
+            age: 26
+        };
+        await colNoEventReduce.insert(insertForSortTest);
+        await colWithEventReduce.insert(insertForSortTest);
+        await testQueryResultForEqualness(
+            colNoEventReduce,
+            colWithEventReduce,
+            queries
+        );
 
         colNoEventReduce.database.destroy();
         colWithEventReduce.database.destroy();